Moving tests files to the source directory
