+++ src/UriTemplate/VarSpecifierTest.php
@@ -11,10 +11,9 @@
 
 declare(strict_types=1);
 
-namespace LeagueTest\Uri\UriTemplate;
+namespace League\Uri\UriTemplate;
 
 use League\Uri\Exceptions\SyntaxError;
-use League\Uri\UriTemplate\VarSpecifier;
 use PHPUnit\Framework\TestCase;
 use function var_export;
