+++ src/UriTemplate/VariableBagTest.php
@@ -11,11 +11,10 @@
 
 declare(strict_types=1);
 
-namespace LeagueTest\Uri\UriTemplate;
+namespace League\Uri\UriTemplate;
 
 use ArrayIterator;
 use League\Uri\Exceptions\TemplateCanNotBeExpanded;
-use League\Uri\UriTemplate\VariableBag;
 use PHPUnit\Framework\TestCase;
 use function var_export;
