+++ src/UriTemplateTest.php
@@ -11,11 +11,10 @@
 
 declare(strict_types=1);
 
-namespace LeagueTest\Uri;
+namespace League\Uri;
 
 use League\Uri\Exceptions\SyntaxError;
 use League\Uri\Exceptions\TemplateCanNotBeExpanded;
-use League\Uri\UriTemplate;
 use PHPUnit\Framework\TestCase;
 
 /**
